+++ headers/Sprite.h
@@ -34,7 +34,7 @@ class Sprite
 	Sprite(SDL_Texture *tex, int width, int height,
 		   unsigned int maxImg);
 	void NextAnim(SpriteControl *ctrl);
-	void PlayAnim(SpriteControl *ctrl, float dt);
+	void PlayAnim(SpriteControl *ctrl, double dt);
 
 	SDL_Rect GetRect(SpriteControl *ctrl);
 	SDL_Texture *GetTexture();
@@ -48,7 +48,7 @@ class Sprite
 	//Number of animation frames
 	unsigned int numFrames_;
 	Direction direction_;
-	float animSpeed_;
+	double animSpeed_;
 
 	int currentFrame_;
 	int animTime_;
